refactor(profiling): move service module to tracer internal
This module will be used to refactor periodic worker and share code between
tracer and profiler.
