+++ ddtrace/profiling/_periodic.py
@@ -3,8 +3,8 @@ import sys
 import threading
 import time
 
+from ddtrace.internal import service
 from ddtrace.profiling import _nogevent
-from ddtrace.profiling import _service
 from ddtrace.vendor import attr
 
 
@@ -154,7 +154,7 @@ def PeriodicRealThread(*args, **kwargs):
 
 
 @attr.s
-class PeriodicService(_service.Service):
+class PeriodicService(service.Service):
     """A service that runs periodically."""
 
     _interval = attr.ib()
